% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{alignReads}
\alias{alignReads}
\title{Read alignment}
\usage{
alignReads(acc)
}
\arguments{
\item{acc}{An accession or a path to an actual SRA file (with .sra suffix)}
}
\value{
aligned reads
}
\description{
This returns the aligned reads.
}
\examples{
alignReads('SRR789392')
alignReads('SRR789334')
}

