% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{alignReadsWithRegion}
\alias{alignReadsWithRegion}
\title{Read alignment}
\usage{
alignReadsWithRegion(acc, refname, start, stop)
}
\arguments{
\item{acc}{An accession or a path to an actual SRA file (with .sra suffix)}

\item{start}{Start position (inclusive)}

\item{stop}{End position (inclusive)}

\item{ref}{The reference name}
}
\value{
aligned reads
}
\description{
This returns the aligned reads.
}
\examples{
alignReadsWithRegion('SRR789392','NC_000020.10', 62926240, 62958722)
alignReadsWithRegion('SRR789334','NC_000020.10', 62926240, 62958722)
}

