% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getFastqReads}
\alias{getFastqReads}
\title{The reads in the read collection.}
\usage{
getFastqReads(acc, max_num_reads)
}
\arguments{
\item{acc}{An accession or a path to an actual SRA file (with .sra suffix)}

\item{n}{The number of reads to return}
}
\value{
the reads in the collection
}
\description{
This returns the all reads.
}
\examples{
getFastqReads('SRR000123',10)
}

