% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getPileUp}
\alias{getPileUp}
\title{The readCount in the read collection.}
\usage{
getPileUp(acc, refname, start = 1L, stop = 0L, MinPileUpDepth = 0L,
  Quality = FALSE)
}
\arguments{
\item{acc}{An accession or a path to an actual SRA file (with .sra suffix)}

\item{refname}{Reference name for pile up}

\item{start}{An in for position of start of pileup}

\item{stop}{An in for position of stop of pileup}

\item{MinPileUpDepth}{Coverage required}
}
\value{
the number of reads in the collection
}
\description{
This simply returns the full read count.
}
\examples{
getPileUp('SRR390728')
}
\author{
Nick Bernstein
}

