% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReads-methods.R
\docType{methods}
\name{getReads,SRR-method}
\alias{getReads,SRR-method}
\title{Get fastq equivalent from SRR record}
\usage{
\S4method{getReads}{SRR}(object, ...)
}
\arguments{
\item{object}{an SRR class object}
}
\value{
a ShortReadQ object
}
\description{
This method will fetch reads from SRA
for the given SRR record. By default, all
reads are returned.
}

