% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getReference}
\alias{getReference}
\title{The readCount in the read collection.}
\usage{
getReference(acc)
}
\arguments{
\item{acc}{An accession or a path to an actual SRA file (with .sra suffix)}
}
\value{
the number of reads in the collection
}
\description{
This simply returns the full read count.
}
\examples{
getReference('SRR390728')
}

