% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQualityvsDepth.R
\name{plotQualityvsDepth.R}
\alias{plotQualityVsDepth}
\alias{plotQualityvsDepth.R}
\title{Plot quality vs depth for an accession}
\usage{
plotQualityVsDepth(acc, ref, start, stop1, depth)
}
\arguments{
\item{acc}{Accension number}

\item{ref}{reference name}

\item{start}{Position start}

\item{stop1}{Position end}

\item{depth}{Depth filter}
}
\details{
stuff
}

