% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{read1}
\alias{read1}
\title{Example of creating a BStringSet in C++}
\usage{
read1(classname, acc, lkup)
}
\arguments{
\item{acc}{An accession or a path to an actual SRA file (with .sra suffix)}

\item{lkup}{lkup = get_seqtype_conversion_lookup('B','DNA')}

\item{n}{The number of reads to return}
}
\value{
the reads in the collection
}
\description{
Example of creating a BStringSet in C++
}
\examples{
require(Biostrings)
lkup = get_seqtype_conversion_lookup('B','DNA')
x = SRA2R:::read1('SRR000123',1000,lkup)
x
}

