% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{reads}
\alias{reads}
\title{The reads in the read collection.}
\usage{
reads(acc, n, lkup)
}
\arguments{
\item{acc}{An accession or a path to an actual SRA file (with .sra suffix)}

\item{n}{The number of reads to return}
}
\value{
the reads in the collection
}
\description{
This simply returns the full read count.
}
\examples{
reads('SRR000123')
}

