% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchSRA.R
\name{searchSRA}
\alias{searchSRA}
\title{Search SRA using Eutils}
\usage{
searchSRA(term)
}
\arguments{
\item{term}{The text used for searching}
}
\value{
A list of SRR ids
}
\description{
Search SRA using Eutils
}
\examples{
ids = searchSRA('breast cancer[WORD] AND cluster_public[prop]')
head(ids)
}

