% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRA2R.R
\docType{package}
\name{SRA2R}
\alias{SRA2R}
\alias{SRA2R-package}
\title{SRA2R: A bridge between the NCBI Sequence Read Archive (SRA) and R}
\description{
The SRA2R package provides three categories of important functions:
}
\section{Convenience functions}{

Install the SRA Toolkit automatically
}

\section{Access SRA data}{

fastq-dump
sam-dump
sra-stat
sra-pileup
}

