% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{SRR-class}
\alias{SRR-class}
\title{An S4 class representing an SRR accession}
\description{
This class simply extends "character" and
checks that the accession matches a regex
for SRR, ERR, or DRR accessions.
}

